refactor: move RPC extension methods into zone-primitives
Move `enabled_tokens` and `encryption_key` from a `ZonePortalExt`
trait in the zone crate to inherent methods on `ZonePortalInstance`
in zone-primitives, gated behind the `rpc` feature. This eliminates
the need for callers to import an extension trait.
