+++ crates/tempo-zone/src/node.rs
@@ -4,7 +4,6 @@
 //! It reuses Tempo's EVM, primitives, and pool, but with noop consensus/network/payload.
 
 use crate::{
-    abi::ZonePortalExt,
     ext::TempoStateExt,
     payload::{ZonePayloadAttributes, ZonePayloadTypes},
 };
